frontend/requirements.txt
aiohttp==3.13.2
python-dotenv==1.2.1
streamlit==1.51.0
